fix: refactor asset registry filtering, mutation, and data management hooks with component support
